power: axp209: Add support for voltage rate control on LDO3
The AXP209 has voltage rate control, or can set a slew rate, for LDO3.
This allows for the power to gradually rise to the desired voltage,
instead of spiking up quickly. Reason to have this can be to reduce the
inrush currents for example.

There are 3 slopes to choose from, the default, 'none' is a voltage rise
of 0.0167 V/uS, a 1.6 mV/uS and a 0.8 mV/uS voltage rise.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>