arm: sunxi: Increase GMAC_TX_DELAY on Lime2
The OLinuXino Lime2 uses the default GMAC_TX_DELAY of 0. After
extensive testing however it turns out that this setting only works for
about 75% of the boards at least with U-Boot 2016.05. Doing analysis on
50+ boards by changing the GMAC_TX_DELAY setting from 0 to 7 a
bell-curve was found with a center-peak at a delay of 3.

It is thus safe to assume that the GMAC_TX_DELAY of 3 should yield the
most reliable result for the Lime2.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>