common/usb_storage: always have usb_stor_scan available
The usb_stor_scan function is currently always has a prototype defined
in the header, but can be undefined in the source.

Further more there is code that relies on this function.

While the proper fix is likely to do this in usb.h, this introduces a
lot more compile errors due to dependencies making this rather trivial
change quite a big one. So lets split is up into two steps. First fix
the code with this trivial change, and leave the reworking of this to
the header for a bigger patch.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>